\name{sabotagdata-package}
\alias{sabotagdata-package}
\alias{sabotagdata}
\docType{package}

\title{
\packageTitle{sabotagdata}
}

\description{
An R-package providing data sets for the Salzburger Botanische Arbeitsgemeinschaft
(SaBotAg).
}

\details{
Nothing here yet.
}

\author{
Maintainer: Roland Kaiser \email{kardinal.eros@gmail.com}
}

\keyword{ package }
\keyword{ spatial }

