\name{BEV}
\alias{BEV}
\docType{data}

\title{
Administrative Boundaries for Austria
}

\description{
Highest precision up-to-date geospatial data for Austria's administrative Boundaries.
}

\usage{data("BEV")}

\references{
\href{https://www.data.gv.at/katalog/dataset/51bdc6dc-25ae-41de-b8f3-938f9056af62}{data.gv.at}
}

\examples{
data(BEV)

names(BEV)
}

\keyword{datasets}

